refactor(engine): split mixer channel strip
